Update basic_commands.tex
ISSUE#19 - DRAFT 1
RND() function description added to Appendix functions section.

Needs review for overall agreement or debate on wording or accuracy (on X16 specifics).

Review code.

Of course, I have Latex output issues as well (due to my inexperience with this system).

LOOKING FOR GENERAL CONCENSUS on whether the wording is okay to commit, even if someone wants to review edit later, is it enough for now?